test(api): add api test suite and reorganize tests
Closes #1160
