+++ backend/tests/Api/SettingsTest.php
@@ -1,16 +1,16 @@
 <?php
 declare(strict_types=1);
 
-namespace Tests\Feature;
+namespace Tests\Api;
 
 use App\Models\Role;
 use App\Models\User;
 use App\Settings\GeneralSettings;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Nuwave\Lighthouse\Testing\MakesGraphQLRequests;
-use Tests\TestCase;
+use Tests\ApiTestCase;
 
-class SettingsTest extends TestCase
+class SettingsTest extends ApiTestCase
 {
     use MakesGraphQLRequests;
     use RefreshDatabase;
